Basic working block editor using ObservableStore
Fork of ObservableStore that allows creating custom Update types for
models. This lets us publish a change description along with state
changes. The controller can use the change description to perform
appropriate updates.
